refactor : 인증 레이아웃 구조 개선 및 스타일 수정
