+++ src/app/onboarding/tech-stack/page.tsx
@@ -1,27 +1,5 @@
-"use client";
-
-import AuthLayout from "@/components/auth/AuthLayout";
-import TechStackSelect from "@/components/auth/TechStackForm";
-import { useUserNickname } from "@/hooks/useUserNickname";
+import TechStackClient from "@/components/auth/TechStackClient";
 
 export default function TechStackPage() {
-  const { nickname, isLoading } = useUserNickname();
-
-  if (isLoading) {
-    return (
-      <div className="flex justify-center items-center min-h-screen">
-        사용자 정보를 불러오는 중...
-      </div>
-    );
-  }
-
-  return (
-    <AuthLayout title="회원가입">
-      {nickname ? (
-        <TechStackSelect nickname={nickname} />
-      ) : (
-        <div>닉네임 정보를 불러올 수 없습니다.</div>
-      )}
-    </AuthLayout>
-  );
+  return <TechStackClient />;
 }
